Add spec to test that WelcomeController#download_csv removes generates csv files

+++ app/controllers/welcome_controller.rb
@@ -57,6 +57,6 @@ class WelcomeController < ApplicationController
   end
 
   def delete_csv
-    File.delete(csv_filename)
+    File.delete(csv_filename) if File.file?(csv_filename)
   end
 end
